+++ src/Payments/Payment.php
@@ -12,9 +12,11 @@ namespace Pronamic\WordPress\Pay\Payments;
 
 use InvalidArgumentException;
 use Pronamic\WordPress\DateTime\DateTime;
+use Pronamic\WordPress\Money\Money;
 use Pronamic\WordPress\Money\TaxedMoney;
 use Pronamic\WordPress\Pay\Address;
 use Pronamic\WordPress\Pay\Customer;
+use Pronamic\WordPress\Pay\MoneyJsonTransformer;
 use Pronamic\WordPress\Pay\Subscriptions\Subscription;
 use Pronamic\WordPress\Pay\Subscriptions\SubscriptionPeriod;
 use Pronamic\WordPress\Pay\TaxedMoneyJsonTransformer;
@@ -40,6 +42,13 @@ class Payment extends LegacyPayment {
 	 * @var TaxedMoney
 	 */
 	private $total_amount;
+
+	/**
+	 * Refunded amount.
+	 *
+	 * @var Money
+	 */
+	private $refunded_amount;
 
 	/**
 	 * The purchase ID.
@@ -318,7 +327,9 @@ class Payment extends LegacyPayment {
 
 		$text = implode( '<br />', $pieces );
 
-		$text = apply_filters( 'pronamic_payment_source_text_' . $this->get_source(), $text, $this );
+		$source = $this->get_source();
+
+		$text = apply_filters( 'pronamic_payment_source_text_' . $source, $text, $this );
 		$text = apply_filters( 'pronamic_payment_source_text', $text, $this );
 
 		return $text;
@@ -369,6 +380,25 @@ class Payment extends LegacyPayment {
 	 */
 	public function set_total_amount( TaxedMoney $total_amount ) {
 		$this->total_amount = $total_amount;
+	}
+
+	/**
+	 * Get refunded amount.
+	 *
+	 * @return Money
+	 */
+	public function get_refunded_amount() {
+		return $this->refunded_amount;
+	}
+
+	/**
+	 * Set refunded amount.
+	 *
+	 * @param Money $refunded_amount Refunded amount.
+	 * @return void
+	 */
+	public function set_refunded_amount( $refunded_amount ) {
+		$this->refunded_amount = $refunded_amount;
 	}
 
 	/**
@@ -656,7 +686,7 @@ class Payment extends LegacyPayment {
 			return null;
 		}
 
-		$this->subscription = new Subscription( $this->subscription_id );
+		$this->subscription = \get_pronamic_subscription( $this->subscription_id );
 
 		return $this->subscription;
 	}
@@ -853,18 +883,28 @@ class Payment extends LegacyPayment {
 
 		$properties = (array) $object;
 
+		// Expiry date.
 		$expiry_date = $this->get_expiry_date();
 
 		if ( null !== $expiry_date ) {
 			$properties['expiry_date'] = $expiry_date->format( \DATE_ATOM );
 		}
 
+		// Total amount.
 		$total_amount = $this->get_total_amount();
 
 		if ( null !== $total_amount ) {
 			$properties['total_amount'] = TaxedMoneyJsonTransformer::to_json( $total_amount );
 		}
 
+		// Refunded amount.
+		$refunded_amount = $this->get_refunded_amount();
+
+		if ( null !== $refunded_amount ) {
+			$properties['refunded_amount'] = MoneyJsonTransformer::to_json( $refunded_amount );
+		}
+
+		// Periods.
 		$periods = $this->get_periods();
 
 		if ( null !== $periods ) {
@@ -873,20 +913,24 @@ class Payment extends LegacyPayment {
 			}
 		}
 
+		// Status.
 		if ( null !== $this->get_status() ) {
 			$properties['status'] = $this->get_status();
 		}
 
+		// Failure reason.
 		$failure_reason = $this->get_failure_reason();
 
 		if ( null !== $failure_reason ) {
 			$properties['failure_reason'] = $failure_reason->get_json();
 		}
 
+		// Google Analytics tracked.
 		if ( null !== $this->get_ga_tracked() ) {
 			$properties['ga_tracked'] = $this->get_ga_tracked();
 		}
 
+		// Origin ID.
 		$origin_id = $this->get_origin_id();
 
 		if ( null !== $origin_id ) {