+++ src/Webhooks/WebhookManager.php
@@ -117,7 +117,7 @@ class WebhookManager {
 
 			/**
 			 * The WordPress Transients API will not always store empty array
-			 * values correctly, therefor we convert an empty array to true.
+			 * values correctly, therefore we convert an empty array to true.
 			 *
 			 * @todo We should probably schedule a daily event to check for
 			 * possible broken webhooks.